Move Netflow processor to cyber-security plugin bug #255
