+++ logisland-docs/processors.rst
@@ -280,42 +280,6 @@ In the list below, the names of required properties appear in **bold**. Any othe
 
 ----------
 
-.. _com.hurence.logisland.processor.bro.NetflowProcessor: 
-
-NetflowProcessor
-----------------
-The `Netflow V5 <http://www.cisco.com/c/en/us/td/docs/ios/solutions_docs/netflow/nfwhite.html>`_ processor is the Logisland entry point to  process Netflow (V5) events. NetFlow is a feature introduced on Cisco routers that provides the ability to collect IP network traffic.We can distinguish 2 components:
-
-	-Flow exporter: aggregates packets into flows and exports flow records (binary format) towards one or more flow collectors
-
-	-Flow collector: responsible for reception, storage and pre-processing of flow data received from a flow exporter
-The collected data are then available for analysis purpose (intrusion detection, traffic analysis...)
-Netflow are sent to kafka in order to be processed by logisland.
-In the tutorial we will simulate Netflow traffic using `nfgen <https://github.com/pazdera/NetFlow-Exporter-Simulator>`_. this traffic will be sent to port 2055. The we rely on nifi to listen of that port for   incoming netflow (V5) traffic and send them to a kafka topic. The Netflow processor could thus treat these events and generate corresponding logisland records. The following processors in the stream can then process the Netflow records generated by this processor.
-
-Class
-_____
-com.hurence.logisland.processor.bro.NetflowProcessor
-
-Tags
-____
-netflow, security
-
-Properties
-__________
-In the list below, the names of required properties appear in **bold**. Any other properties (not in bold) are considered optional. The table also indicates any default values
-.
-
-.. csv-table:: allowable-values
-   :header: "Name","Description","Allowable Values","Default Value","Sensitive","EL"
-   :widths: 20,60,30,20,10,10
-
-   "debug", "Enable debug. If enabled, the original JSON string is embedded in the record_value field of the record.", "", "null", "", ""
-   "output.record.type", "the output type of the record", "", "netflowevent", "", ""
-   "enrich.record", "Enrich data. If enabledthe netflow record is enriched with inferred data", "", "false", "", ""
-
-----------
-
 .. _com.hurence.logisland.processor.NormalizeFields: 
 
 NormalizeFields
@@ -457,6 +421,69 @@ com.hurence.logisland.processor.bro.ParseBroEvent
 Tags
 ____
 bro, security, IDS, NIDS
+
+Properties
+__________
+In the list below, the names of required properties appear in **bold**. Any other properties (not in bold) are considered optional. The table also indicates any default values
+.
+
+.. csv-table:: allowable-values
+   :header: "Name","Description","Allowable Values","Default Value","Sensitive","EL"
+   :widths: 20,60,30,20,10,10
+
+   "debug", "Enable debug. If enabled, the original JSON string is embedded in the record_value field of the record.", "", "null", "", ""
+
+----------
+
+.. _com.hurence.logisland.processor.netflow.ParseNetflowEvent: 
+
+ParseNetflowEvent
+-----------------
+The `Netflow V5 <http://www.cisco.com/c/en/us/td/docs/ios/solutions_docs/netflow/nfwhite.html>`_ processor is the Logisland entry point to  process Netflow (V5) events. NetFlow is a feature introduced on Cisco routers that provides the ability to collect IP network traffic.We can distinguish 2 components:
+
+	-Flow exporter: aggregates packets into flows and exports flow records (binary format) towards one or more flow collectors
+
+	-Flow collector: responsible for reception, storage and pre-processing of flow data received from a flow exporter
+The collected data are then available for analysis purpose (intrusion detection, traffic analysis...)
+Netflow are sent to kafka in order to be processed by logisland.
+In the tutorial we will simulate Netflow traffic using `nfgen <https://github.com/pazdera/NetFlow-Exporter-Simulator>`_. this traffic will be sent to port 2055. The we rely on nifi to listen of that port for   incoming netflow (V5) traffic and send them to a kafka topic. The Netflow processor could thus treat these events and generate corresponding logisland records. The following processors in the stream can then process the Netflow records generated by this processor.
+
+Class
+_____
+com.hurence.logisland.processor.netflow.ParseNetflowEvent
+
+Tags
+____
+netflow, security
+
+Properties
+__________
+In the list below, the names of required properties appear in **bold**. Any other properties (not in bold) are considered optional. The table also indicates any default values
+.
+
+.. csv-table:: allowable-values
+   :header: "Name","Description","Allowable Values","Default Value","Sensitive","EL"
+   :widths: 20,60,30,20,10,10
+
+   "debug", "Enable debug. If enabled, the original JSON string is embedded in the record_value field of the record.", "", "null", "", ""
+   "output.record.type", "the output type of the record", "", "netflowevent", "", ""
+   "enrich.record", "Enrich data. If enabledthe netflow record is enriched with inferred data", "", "false", "", ""
+
+----------
+
+.. _com.hurence.logisland.processor.pcap.ParsePCap: 
+
+ParsePCap
+---------
+The PCap processor is the LogIsland entry point to parse network packets captured in pcap format. The PCap processor decodes the bytes of the incoming pcap record, where a Global pcap header followed by a sequence of [packet header, packet data] pairs are stored. Each incoming pcap event is parsed into n packet records where packet headers data is made available in dedicated fields.
+
+Class
+_____
+com.hurence.logisland.processor.pcap.ParsePCap
+
+Tags
+____
+PCap, security, IDS, NIDS
 
 Properties
 __________
