+++ logisland-framework/logisland-resources/src/main/resources/conf/index-netflow-events.yml
@@ -1,0 +1,103 @@
+#########################################################################################################
+# Logisland configuration script example: PArse Netflow events and index them into ElasticSearch
+#########################################################################################################
+
+version: 0.9.8
+documentation: LogIsland analytics main config file. Put here every engine or component config
+
+#########################################################################################################
+# engine
+engine:
+  component: com.hurence.logisland.engine.spark.KafkaStreamProcessingEngine
+  type: engine
+  documentation: Index Netflow events with LogIsland
+  configuration:
+    spark.app.name: IndexNetFlowEventsDemo
+    spark.master: local[4]
+    spark.driver.memory: 1G
+    spark.driver.cores: 1
+    spark.executor.memory: 2G
+    spark.executor.instances: 4
+    spark.executor.cores: 2
+    spark.yarn.queue: default
+    spark.yarn.maxAppAttempts: 4
+    spark.yarn.am.attemptFailuresValidityInterval: 1h
+    spark.yarn.max.executor.failures: 20
+    spark.yarn.executor.failuresValidityInterval: 1h
+    spark.task.maxFailures: 8
+    spark.serializer: org.apache.spark.serializer.KryoSerializer
+    spark.streaming.batchDuration: 4000
+    spark.streaming.backpressure.enabled: false
+    spark.streaming.unpersist: false
+    spark.streaming.blockInterval: 500
+    spark.streaming.kafka.maxRatePerPartition: 3000
+    spark.streaming.timeout: -1
+    spark.streaming.unpersist: false
+    spark.streaming.kafka.maxRetries: 3
+    spark.streaming.ui.retainedBatches: 200
+    spark.streaming.receiver.writeAheadLog.enable: false
+    spark.ui.port: 4050
+  streamConfigurations:
+
+
+    # Parsing
+    - stream: parsing_stream
+      component: com.hurence.logisland.stream.spark.KafkaRecordStreamParallelProcessing
+      type: stream
+      documentation: A processor chain that transforms Netflow events into Logisland records
+      configuration:
+        kafka.input.topics: netflow
+        kafka.output.topics: logisland_events
+        kafka.error.topics: logisland_errors
+        kafka.input.topics.serializer: none
+        kafka.output.topics.serializer: com.hurence.logisland.serializer.KryoSerializer
+        kafka.error.topics.serializer: com.hurence.logisland.serializer.JsonSerializer
+        kafka.metadata.broker.list: sandbox:9092
+        kafka.zookeeper.quorum: sandbox:2181
+        kafka.topic.autoCreate: true
+        kafka.topic.default.partitions: 2
+        kafka.topic.default.replicationFactor: 1
+      processorConfigurations:
+
+        # Transform Netflow events into Logisland records
+        - processor: Netflow adaptor
+          component: com.hurence.logisland.processor.netflow.ParseNetflowEvent
+          type: parser
+          documentation: A processor that transforms Netflow events into LogIsland events
+          configuration:
+            debug: false
+            enrich.record: false
+
+    # Indexing
+    - stream: indexing_stream
+      component: com.hurence.logisland.stream.spark.KafkaRecordStreamParallelProcessing
+      type: processor
+      documentation: A processor chain that pushes Netflow events to ES
+      configuration:
+        kafka.input.topics: logisland_events
+        kafka.output.topics: none
+        kafka.error.topics: logisland_errors
+        kafka.input.topics.serializer: com.hurence.logisland.serializer.KryoSerializer
+        kafka.output.topics.serializer: none
+        kafka.error.topics.serializer: com.hurence.logisland.serializer.JsonSerializer
+        kafka.metadata.broker.list: sandbox:9092
+        kafka.zookeeper.quorum: sandbox:2181
+        kafka.topic.autoCreate: true
+        kafka.topic.default.partitions: 2
+        kafka.topic.default.replicationFactor: 1
+      processorConfigurations:
+
+        # Put into ElasticSearch
+        - processor: ES Publisher
+          component: com.hurence.logisland.processor.elasticsearch.PutElasticsearch
+          type: processor
+          documentation: A processor that pushes Netflow events into ES
+          configuration:
+            default.index: netflow
+            default.type: events
+            hosts: sandbox:9300
+            cluster.name: elasticsearch
+            batch.size: 2000
+            timebased.index: today
+            es.index.field: search_index
+            es.type.field: record_type
