+++ logisland-plugins/logisland-cyber-security-plugin/src/main/java/com/hurence/logisland/processor/netflow/ParseNetflowEvent.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hurence.logisland.processor.bro;
+package com.hurence.logisland.processor.netflow;
 
 import com.hurence.logisland.annotation.documentation.CapabilityDescription;
 import com.hurence.logisland.annotation.documentation.Tags;
@@ -52,9 +52,9 @@ import java.util.*;
                 + "in the stream can then process the Netflow records generated by this processor.")
 
 
-public class NetflowProcessor extends AbstractProcessor {
+public class ParseNetflowEvent extends AbstractProcessor {
 
-    private static Logger logger = LoggerFactory.getLogger(NetflowProcessor.class);
+    private static Logger logger = LoggerFactory.getLogger(ParseNetflowEvent.class);
 
     final int FIN = 1;
     final int SYN = 2;
@@ -122,9 +122,7 @@ public class NetflowProcessor extends AbstractProcessor {
          */
         records.forEach(record -> {
             byte[] recordValue=null;
-            byte[] recordValue1=null;
             try {
-                recordValue1 = ((String) record.getField(FieldDictionary.RECORD_VALUE).getRawValue()).getBytes("US-ASCII");
                 recordValue = ((String) record.getField(FieldDictionary.RECORD_VALUE).getRawValue()).getBytes("ISO-8859-1");
             } catch (UnsupportedEncodingException e) {
                 logger.error("Error while retrieving the record value : record skipped");
